% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmhclust.R
\name{gmhclust}
\alias{gmhclust}
\title{Run Hierararchical Mahalanobis clustering accelerated on GPU}
\usage{
gmhclust(
  dataMatrix,
  threshold = 0.5,
  subthreshHandling = "mahal",
  normalize = FALSE
)
}
\arguments{
\item{dataMatrix}{Real matrix of cells in column major order (each column one cell)}

\item{threshold}{Real number in the interval of (0,1) defining
the minimal relative size of cluster (relative to the number of
observations) whose distance to other clusters will be computed as a
pure Mahalanobis distance.}

\item{subthreshHandling}{One of string values ('mahal','mahal0','euclid','euclidMahal'); 
it is a method how subthreshold clusters distort the space around them - i.e. how}

\item{normalize}{Boolean; if TRUE, cluster size
will be ignored when computing Mahalanobis distance from the cluster.}
}
\value{
Returns list(merge, height);
merge is the (n-1) x 2 matrix describing the iterative merging of observations and clusters (column-major style)
height is the 1 x (n-1) matrix that stores the heights of the n-1 clusters
}
\description{
Run Hierararchical Mahalanobis clustering accelerated on GPU
}
